refactor: moved videoReducer to a new file
